ML/requirements.txt
flask==3.0.3
requests==2.32.3
pandas==2.2.3
mlxtend==0.23.1
python-dotenv==1.0.1

numpy
scikit-learn
tensorflow